+++ CHANGELOG.md
@@ -2,21 +2,29 @@
 
 ## [UNRELEASED]
 
+### Added
+
+- Install script
+
 ## [0.2.0] - 2024-08-17
 
 ### Added
+
 - Service file in release
 
 ### Changed
+
 - Release tar archive gzip
-- Adapt release content structure 
+- Adapt release content structure
 
 ## [0.1.0] - 2024-07-06
 
 ### Added
+
 - Initial Version
 
-
 [0.2.0]: https://github.com/EdwarDDay/kotlin-server-scripts/releases/tag/0.2.0
+
 [0.1.0]: https://github.com/EdwarDDay/kotlin-server-scripts/releases/tag/0.1.0
+
 [UNRELEASED]: https://github.com/EdwarDDay/kotlin-server-scripts